The old tests for Workpattern are now working using a Time object internally.
Now all my tests are passing again, but I need more tests because
now I am allowing users to pass Date, DateTime and Time objects
with their own Timezone and getthe right answer to include daylight
saving time or not.

I live in the UK and we are on GMT/UTC as I write this so I have
changed my computer to be east coast us (and all my old UTC tests pass).

I need to make sure I write tests that can ran in any timezone and make use
of more than one Timezone and also Date DateTime and Time objects.  This
is only needed for Workpattern because everything else is in utc Time.

Feeling confident but still a long way to go to prove it.