One thing that can happen when you come back to code after many weeks or months is you have forgotten what every thing does and if you only occasionaly look at that code in that time then you leave things half finished.
Here I was reviewing where I was in changing to Timezones and guess what?
I realised the Day class was made redundant ages ago and I hadn't removed it
and the tests that came with it from what I was doing.

This commit just made Day class and tests history.