simplified duration calculations in Day
The method #working_minutes can calculate the working minutes in any
segment of a single day given the start and finish times.

This can make a lot of complicated methods redundant in Week class and
means the calculation of time in a day is the responsibility of the Day
class.