all tests passing again BUT ...
There are a couple of places where the tests take too long and I
need to look at improving performance.

There are some tests which just assert true and they need to be
written out.

There are some cases where the fault was not in the class being
tested so I need to put the appropriate test in the failed class.
This is to catch it if it reocurrs