ignore sorted_set for ruby v2.3.x and below
ruby v2.3.x and below use the SortedSet which came with ruby but this
was removed from ruby v3.0.x and put in the sorted_set gem.  The ruby
versions below v2.4  are not able to use it and this change ensures
sorted_set is only required for the right versions.