changed to match bundle gem name
I created a gem using `bundle gem mygem` to see what the more recent
approach is as I created Workpattern in 2009 using gems that have been
superceded and may be doing the wrong thing.

The `.gitignore` file was updated here and a number of lines removed.
The only other change was to stop ignoring Gemfile.lock.