upgraded minitest and rake
This was to match what was in `bundle gem mygem`

The minitest version went downand this introduced a bunch of Warnings,
which it seems are something to do with encoding.  They were removed by
overwriting from the start of the line to the ruby reserved word or
words.

This also showed a warning about an else statement with a test
expression afterwards.  Fixed by removing the test expression.