removed ruby v1.9.3 from travis-ci
Travis-ci fails the  ruby v1.9.3 build of wrokpattern despite tests
passing on my machines with the same versions of everything.

I don't understand it, but I do know it leaves a failing build on my
repository.  The ruby v1.9.3 has been out of support since about 2015 so
I am going to deprecate it.  I am removing it from the travis-ci build
because I know it is fine and this is clearly an issue with my knowledge
of travis.

I willl continue to test on v1.9.3 but not as part of travis.